Move bulk of behavior into StandaloneView module
